Criaçao das Classes AreaDeConhecimento e AreaDeConhecimentoDAO ajustes relacionados às inclusões. remoção de bibliotecas desnecessárias.

+++ Produto/Implementacao/Sistema-Evento/src/java/persistencia/CategoriaDAO.java
@@ -16,7 +16,7 @@ public class CategoriaDAO {
         sessao = HibernateUtil.getSessionFactory().openSession();
     }
     
-    public Categoria carregar(Long id) {
+    public Categoria carregar(int id) {
         return (Categoria) sessao.get(Categoria.class, id);
     }
